Move offset manager tests to a test folder
